servicedb/requirements.txt
aio-pika==8.2.5
alembic==1.8.1
fastapi==0.87.0
FastAPI-SQLAlchemy==0.2.1
psycopg2-binary==2.9.5
uvicorn==0.20.0